Added basic ssl support.
Requires gnutls.

+++ src/lib/typedefs.h
@@ -30,7 +30,7 @@
 typedef unsigned char      byte;
 typedef unsigned char      uint8;
 typedef unsigned short     uint16;
-typedef unsigned int       uint32;
+typedef unsigned long      uint32;
 typedef int                int32;
 typedef unsigned long long uint64;
